standardizing first and group level
